Seventh Commit --- Still building basics Working on Newsletter
Implicated following pages:
- start
- signIn
- signUp
- homepage
- Quiz
- Quiz Component
- Newsletter
